Fix uncaught type discrepancy
Fix type difference that implicitly casted to not matter.
Specifically, even though cast the C intrinsic from a `long`
to an `unsigned int`, accidentally decalred `maxTargetI` as
a `long` for some reason.

+++ src/algorithms/libs/volk_gnsssdr_module/volk_gnsssdr/kernels/volk_gnsssdr/volk_gnsssdr_8i_index_max_16u.h
@@ -660,7 +660,7 @@ static inline void volk_gnsssdr_8i_index_max_16u_rvv(unsigned int* target, const
                     */
 
                     // maxTargetI = index of first set bit in maxTarget
-                    long maxTargetI = (unsigned int) __riscv_vfirst_m_b1(maxTargetMask, vl);
+                    unsigned int maxTargetI = (unsigned int) __riscv_vfirst_m_b1(maxTargetMask, vl);
                     // Cast is risky; keep eye out
 
                     unsigned int elapsedN = num_points - n;
